show warning message when user change setting endpoint url if there is an unsubmitted scorecard
