+++ app/localization/en.json
@@ -238,5 +238,7 @@
   "previous": "Previous",
   "skip": "Skip",
   "attributes": "Attributes",
-  "scorecardSubmitFailed": "Scorecard submit failed"
+  "scorecardSubmitFailed": "Scorecard submit failed",
+  "notice": "Notice",
+  "scorecardRemainingMessage": "There are remaining implemented scorecards and not submit yet, please submit them first"
 }
