+++ app/localization/km.json
@@ -199,9 +199,9 @@
   "ok": "យល់ព្រម",
   "version": "កំណែ",
   "didYouEnterTheUrlCorrectly": "តើអ្នកបានបញ្ចូល URL ត្រឹមត្រូវទេ?",
-  "ifYouKeepHavingThisProblem": "ប្រសិនបើអ្នកនៅតែជួបប្រទះបញ្ហានេះសូមរាយការណ៍ទៅកាន់អ្នកដែលស្នើរសុំឱ្យអ្នកប្រមូលទិន្នន័យ។",
+  "ifYouKeepHavingThisProblem": "ប្រសិនបើអ្នកនៅតែជួបប្រទះបញ្ហានេះសូមរាយការណ៍ទៅកាន់អ្នកដែលស្នើសុំឱ្យអ្នកប្រមូលទិន្នន័យ។",
   "pleaseAddCriteria": "សូម​ជ្រើស​រើស​លក្ខណៈវិនិច្ឆ័យ",
-  "raisedTimes": "ចំនួនដងដែលបានស្នើរឡើង",
+  "raisedTimes": "ចំនួនដងដែលបានស្នើឡើង",
   "criteriaList": "បញ្ជីលក្ខណៈវិនិច្ឆ័យ",
   "pleaseSelectCriteria": "សូមធ្វើកាជ្រើសរើសលក្ខណៈវិនិច្ឆ័យ",
   "numberOfSelectedCriterias": "ចំនួនលក្ខណៈវិនិច្ឆ័យដែលបានជ្រើសរើស",
@@ -238,5 +238,7 @@
   "previous": "ត្រលប់ក្រោយ",
   "skip": "រំលង",
   "attributes": "លក្ខណៈសម្បត្តិ",
-  "scorecardSubmitFailed": "បរាជ័យក្នុងការដាក់ស្នើរប័ណ្ណដាក់ពិន្ទុ"
+  "scorecardSubmitFailed": "បរាជ័យក្នុងការដាក់ស្នើប័ណ្ណដាក់ពិន្ទុ",
+  "notice": "សម្គាល់",
+  "scorecardRemainingMessage": "នៅមានប័ណ្ណដាក់ពិន្ទុដែលបានអនុវត្តហើយតែមិនទាន់បានធ្វើការដាក់ស្នើ សូមធ្វើការដាក់ស្នើប័ណ្ណដាក់ពិន្ទុទាំងនោះជាមុន"
 }
